Lab_08/moisure-writer/requirements.txt
kafka-python==2.0.2
rasterio==1.3.6